% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiassumptions.R
\name{chiassumptions}
\alias{chiassumptions}
\title{Chi Assumptions}
\usage{
chiassumptions(dataset)
}
\arguments{
\item{dataset}{The data set you are testing}
}
\value{
The assumptions for the Chi Squared test.
}
\description{
Prints the assumptions for the Chi Squared test.
}
\examples{
chiassumptions(project)
}
