% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chidecision.R
\name{chidecision}
\alias{chidecision}
\title{Title}
\usage{
chidecision(dataset)
}
\arguments{
\item{chifit1}{The list of output created by the chifit function}
}
\value{

}
\description{
Chooses whether or not to reject the null hypothesis based upon the p-value of the chi squared test.
}
