% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chifit.R
\name{chifit}
\alias{chifit}
\title{Chi Fit}
\usage{
chifit(dataset)
}
\arguments{
\item{dataset}{The data set you are testing}
}
\value{

}
\description{
Applies the chi squared test and returns relevant ouput in a list.
}
