% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chihypothesis.R
\name{chihypothesis}
\alias{chihypothesis}
\title{Chi Hypotheses}
\usage{
chihypothesis(dataset)
}
\arguments{
\item{dataset}{The data set you are testing}
}
\value{
The hypothesis for the Chi Squared test.
}
\description{
Prints the null and alternative hypotheses for the Chi Squared test.
}
\examples{
chihypothesis(project)
}
