% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_dataset.R
\docType{data}
\name{error_dataset}
\alias{error_dataset}
\title{Data set to test errors.}
\format{
A data frame with 21 rows and 1 variables:
\describe{
  \item{height}{height of the person, in cm}
}
}
\usage{
error_dataset
}
\description{
Data set used to test errors of statistical functions with testthat().
}
\keyword{datasets}
