% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmassumptions.R
\name{lmassumptions}
\alias{lmassumptions}
\title{Lm Assumptions}
\usage{
lmassumptions(dataset)
}
\arguments{
\item{dataset}{The data set you are testing}
}
\value{

}
\description{
Prints the assumptions for the linear model test.
}
\examples{
lmassumptions(project)
}
