% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmconclusion.R
\name{lmconclusion}
\alias{lmconclusion}
\title{Lm Conclusion}
\usage{
lmconclusion(dataset)
}
\arguments{
\item{dataset}{The data set you are testing}
}
\value{

}
\description{
Gives a conclusion for the linear model test based upon the p-value obtained.
}
\examples{
lmconclusion(project)
}
