% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmdecision.R
\name{lmdecision}
\alias{lmdecision}
\title{Lm decision}
\usage{
lmdecision(dataset)
}
\arguments{
\item{dataset}{The data set you are testing}
}
\value{

}
\description{
Chooses whether or not to reject the null hypothesis based upon the p-value of the linear model test.
}
\examples{
lmdecision(project)
}
