% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmfit.R
\name{lmfit}
\alias{lmfit}
\title{Lm fit}
\usage{
lmfit(file_name)
}
\arguments{
\item{dataset}{The data set you are testing}
}
\value{

}
\description{
Applies the linear model test and returns relevant ouput in a list.
}
