% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmhypothesis.R
\name{lmhypothesis}
\alias{lmhypothesis}
\title{Lm Hypotheses}
\usage{
lmhypothesis(dataset)
}
\arguments{
\item{dataset}{The data set you are testing}
}
\value{
The hypotheses for the linear model test.
}
\description{
Prints the null and alternative hypotheses for the linear model test.
}
\examples{
lmhypothesis(project)
}
