% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylmtest.R
\name{mylmtest}
\alias{mylmtest}
\title{Full Lm Test}
\usage{
mylmtest(dataset)
}
\arguments{
\item{dataset}{The data set you wish to test}
}
\value{

}
\description{
Provides a full linear model test, with hypotheses, assumptions, a decision and a conclusion.
}
