% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.R
\name{myttest}
\alias{myttest}
\title{Full T-Test}
\usage{
myttest(dataset)
}
\arguments{
\item{dataset}{The data set you wish to test}
}
\value{

}
\description{
Provides a full t-test, with hypotheses, assumptions, a decision and a conclusion.
}
\examples{
myttest(project)
}
