% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tconclusion.R
\name{tconclusion}
\alias{tconclusion}
\title{T Conclusion}
\usage{
tconclusion(dataset)
}
\arguments{
\item{dataset}{The data set you are testing}
}
\value{
Conclusion for the t-test.
}
\description{
Gives a conclusion for the t-test based upon the p-value obtained.
}
\examples{
tconclusion(project)
}
