% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdecision.R
\name{tdecision}
\alias{tdecision}
\title{T decision}
\usage{
tdecision(dataset)
}
\arguments{
\item{dataset}{The data set you are testing}
}
\value{

}
\description{
Chooses whether or not to reject the null hypothesis based upon the p-value of the t-test.
}
\examples{
tdecision(project)
}
