% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfit.R
\name{tfit}
\alias{tfit}
\title{T fit}
\usage{
tfit(dataset)
}
\arguments{
\item{dataset}{The data set you are testing}
}
\value{
Relevent t-test statistics.
}
\description{
Applies the t-test and returns relevant ouput in a list.
}
\examples{
tfit(project)
}
