% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiconclusion.R
\name{chiconclusion}
\alias{chiconclusion}
\title{Chi Conclusion}
\usage{
chiconclusion(dataset)
}
\arguments{
\item{dataset}{The data set you are testing}
}
\value{

}
\description{
Gives a conclusion for the chi squared test based upon the p-value obtained.
}
