% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chidecision.R
\name{chidecision}
\alias{chidecision}
\title{Chi decision}
\usage{
chidecision(dataset)
}
\arguments{
\item{dataset}{The data set you are testing}
}
\value{

}
\description{
Chooses whether or not to reject the null hypothesis based upon the p-value of the chi squared test.
}
