% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mychitest.R
\name{mychitest}
\alias{mychitest}
\title{Full Chi Squared Test}
\usage{
mychitest(dataset)
}
\arguments{
\item{dataset}{The data set you wish to test}
}
\value{
Full Chi Squared test report.
}
\description{
Provides a full chi squared test, with hypotheses, assumptions, a decision and a conclusion.
}
\examples{
mychitest(project)
}
