% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tassumptions.R
\name{tassumptions}
\alias{tassumptions}
\title{T Assumptions}
\usage{
tassumptions(dataset)
}
\arguments{
\item{dataset}{The data set you are testing}
}
\value{

}
\description{
Prints the assumptions for the t-test.
}
\examples{
tassumptions(project)
}
