% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thypothesis.R
\name{thypothesis}
\alias{thypothesis}
\title{T Hypotheses}
\usage{
thypothesis(dataset)
}
\arguments{
\item{dataset}{The data set you are testing}
}
\value{

}
\description{
Prints the null and alternative hypotheses for the t-test.
}
